requirements.txt
numpy
tqdm
pyyaml
soundfile
librosa
opencv-python

